requirements.txt
fastmcp
pydantic
starlette
boto3
botocore
tavily-python